* Frage und jeweiliger Typ Done * Gruppe Done * Antwort Möglichkeiten Done

+++ src/main/java/com/htw/test/Links.java
@@ -2,7 +2,7 @@ package com.htw.test;
 
 public class Links {
 	
-	public static int calculateCheckDigit(String numString) {
+	public static long calculateCheckDigit(String numString) {
 		
 		long checksum = 0;
 		
@@ -11,23 +11,23 @@ public class Links {
 			checksum += (charValue - 48) * ( i % 2 > 0 ? 2 : 1 );
 		}
 		
-		return (int) ( 10 - checksum % 10 ) % 10;
+		return (long) ( 10 - checksum % 10 ) % 10;
 		
 	}
 
-	public static String createLink(int id) {
+	public static String createLink(long id) {
 		
-		Long sourceId = ((long) id) * 5 + 104302;
+		long sourceId = id * 5 + 104302;
 		String numString = Long.toString( sourceId );
-		int checkDigit = calculateCheckDigit(numString);
+		long checkDigit = calculateCheckDigit(numString);
 		return numString + String.valueOf(checkDigit);
 		
 	}
 	
-	public static int decodeLink(String link) {
+	public static long decodeLink(String link) {
 		
 		String numString = link.substring(0, link.length() - 1);
-		int id = (int) ( (Long.parseLong(numString) - 104302) / 5 );
+		long id = (long) ( (Long.parseLong(numString) - 104302) / 5 );
 		return createLink(id).equals(link) ? id : 0;
 		
 	}
