+++ src/main/java/com/htw/test/repositories/AntwortRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.htw.test.model.Antworten;
 
-public interface AntwortRepository extends CrudRepository<Antworten, Integer> {
+public interface AntwortRepository extends CrudRepository<Antworten, Long> {
 
 }
