+++ src/main/java/com/htw/test/repositories/FrageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.htw.test.model.Frage;
 
-public interface FrageRepository extends CrudRepository<Frage, Integer> {
+public interface FrageRepository extends CrudRepository<Frage, Long> {
 
 }
