+++ src/main/java/com/htw/test/repositories/GruppeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.htw.test.model.Gruppe;
 
-public interface GruppeRepository extends CrudRepository<Gruppe, Integer>{
+public interface GruppeRepository extends CrudRepository<Gruppe, Long>{
 
 }
